refactor(archive): own semantic pricing
