+++ docs/topology-status.md
@@ -22,7 +22,7 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 | archive-provider | — | 3 | 3 | 0 | 0 |
 | archive-query | archive query semantics | 17 | 17 | 0 | 0 |
 | archive-raw-payload | — | 5 | 5 | 0 | 0 |
-| archive-semantic | — | 5 | 5 | 0 | 0 |
+| archive-semantic | — | 6 | 6 | 0 | 0 |
 | archive-session | — | 6 | 6 | 0 | 0 |
 | archive-viewport | — | 5 | 5 | 0 | 0 |
 | core-primitive | — | 9 | 9 | 0 | 0 |
@@ -31,7 +31,7 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 
 - **Stable** (no move scoped): 582
 - **Kernel** (polylogue/ root): 10
-- **Primitives** (lib-root + storage-root): 15
+- **Primitives** (lib-root + storage-root): 14
 - **TBD** (cell needs explicit assignment): 0
 - **Total declared**: 695
 - **Realized polylogue/**/*.py**: 695 files declared
